pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "florence"
version = "0.5.0"
description = "Healthcare-oriented document reading MVP with persistent storage, QA and answer rewriting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.0",
    "fastapi>=0.115.0",
    "pypdf>=6.8.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "pytest>=9.0.2",
    "sentence-transformers>=5.1.1",
    "sentencepiece>=0.2.0",
    "torch>=2.2.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["app"]